requirements.txt
# Core
numpy==1.26.4
tensorflow==2.21.0
tf_keras==2.21.0
keras==3.12.1

# Quantization
tensorflow-model-optimization==0.8.0

# ECG / Signal
wfdb==4.3.1
PyWavelets==1.8.0
mat73==0.65

# ML utilities
scikit-learn==1.7.2
imbalanced-learn==0.14.1
scipy==1.15.3

# Visualization & logging
matplotlib==3.10.8
seaborn==0.13.2
tensorboardX==2.6.5
Pillow==12.2.0

# Excel export
openpyxl==3.1.5
